requirements.txt
tensorflow>=2.12.0
opencv-python>=4.7.0
numpy>=1.23.0
sympy>=1.11.0
scikit-learn>=1.2.0
Pillow>=9.0.0
